Add exclusion for __main__.py for certain errors
The way Plac annotations work, it's much more readable to put extra
spaces in the list in order to align the columns.  Unfortunately,
py.test doesn't have a way to ignore errors in a block of code, and
it's too messy to add per-line exclusions on every line of the Plac
annotation.  The best we can do is ignore the codes for the file,
which is not a great solution, but the best I can think of.

See also this Stack Overflow answer by the py.test maintainer:
https://stackoverflow.com/a/64431741/743730